refactor: renames src/output to src/markdown
